Started working on better Animations

+++ SpriteSheet.cpp
@@ -95,21 +95,16 @@ void SpriteSheet::Draw() {
 	
 }
 
-void SpriteSheet::Draw(int index, int x, int y) {
+void SpriteSheet::Draw(int index, float x, float y) {
 		
 	D2D_RECT_F src = D2D1::RectF(
-		static_cast<float>((index % spritesAccross) * spriteWidth),		// claculate position of sprite
-		static_cast<float>((index / spritesAccross) * spriteHeight),
-		static_cast<float>((index % spritesAccross) * spriteHeight) + spriteWidth,	//calculate size of sprite
-		static_cast<float>((index / spritesAccross) * spriteHeight) + spriteHeight
+		(index % spritesAccross) * spriteWidth,		// claculate position of sprite
+		(index / spritesAccross) * spriteHeight,
+		(index % spritesAccross) * spriteHeight + spriteWidth,	//calculate size of sprite
+		(index / spritesAccross) * spriteHeight + spriteHeight
 	);
 		
-	D2D_RECT_F dest = D2D1::RectF(
-		static_cast<float>(x),
-		static_cast<float>(y),
-		static_cast<float>(x + spriteWidth), 
-		static_cast<float>(y + spriteHeight)
-	);
+	D2D_RECT_F dest = D2D1::RectF( x, y, x + spriteWidth, y + spriteHeight);
 
 
 	gfx->GetRenderTarget()->DrawBitmap(
